requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
apscheduler==3.10.4
icalendar==5.0.13
python-dateutil==2.9.0
pytz==2024.1
openpyxl==3.1.5
